changed some of the int operations to cast to double first in attept to get the subpixel animation working
git-svn-id: http://svn.code.sf.net/p/sonia/source_code@172 fb298286-8e17-4df1-bf80-b634edc3145b

+++ trunk/sonia/RenderSlice.java
@@ -141,8 +141,8 @@ public class RenderSlice {
 		double[] xCoords = layoutEngine.getCurrentXCoords();
 		double[] yCoords = layoutEngine.getCurrentYCoords();
 		// how for to translate the coords by
-		int left = layoutEngine.getLeftPad();
-		int top = layoutEngine.getTopPad();
+		double left = (double)layoutEngine.getLeftPad();
+		double top = (double)layoutEngine.getTopPad();
 
 		// Color origColor = graphics.getColor();
 		// Font originalFont = graphics.getFont();
